Read in user input from stdin
While stdin is valid, read text input char by char from it until EOF
reached. Store characters in a std::string and echo it to stdout at
end of program.

Update README to describe how to input text, and to end input by Return
and CTRL-D to give End-Of-File( (EOF).