Implement -i/-o command line options
Provide -i and -o command line options for taking input and output files
for reading/writing text from/to. Expect these to be followed by the
name of the file, but only check that this name exists, not that it
is a valid file.

Update help output and README to document these options, noting that
they are not yet fully implemented.