Provide header for transformChar interface
Move declaration of `transformChar` function into dedicated header under
MPAGSCipher directory. Replace declaration in main mpags-cipher file
with inclusion of this header.

Add header to build of mpags-cipher to ensure it appears in CMake
generated IDE projects. Declare MPAGSCipher as an include directory
of mpags-cipher to ensure the header is found by the build.