Implement read/write of input/output from/to file
If supplied, use input(output) file name option arguments to read(write)
input(output) text from(to) a file stream created on the file. If file
stream is not `good` emit an error and exit program.

Fallback to reading/writing from/to stdin/out when filenames are not
supplied.

Update README to indicate support for file I/O.