Provide minimal implementation of CaesarCipher
Add header/source for CaesarCipher function taking string to be
encrypted/decrypted, the shift key as a string, a boolean to
indicate if decryption is required and a reference for the output
string to be filled.

Conversion of the string key to an integer shift is done via std::stoul
bu exceptions are not handled.

Call CaesarCipher in main program, using inputs from command line to
configure.